feat: sidebar animated with mui
